[ENH] big reorg to move things from main.py to features.py or regression.py
